+++ lib/src/core/core.dart
@@ -1,4 +1,5 @@
 export 'errors.dart';
+export 'services/language_validation_service.dart';
 export 'localization_generator.dart';
 export 'logger.dart';
 export 'string_utils.dart';